Rename files and Views
