+++ AzooKeyCore/Sources/KeyboardViews/ApplicationSpecificKeyboardViewSettingProvider.swift
@@ -30,7 +30,7 @@ import Foundation
     static var enablePasteButton: Bool { get }
     static var displayTabBarButton: Bool { get }
     static var hideResetButtonInOneHandedMode: Bool { get }
-    static var useBetaMoveCursorBar: Bool { get }
+    static var useSliderStyleCursorBar: Bool { get }
     static var useShiftKey: Bool { get }
 
     static var canResetLearningForCandidate: Bool { get }
